feat: shape page를 fsd 패턴으로 나눔
