+++ src/app/info/shape/components/contentBox.tsx
@@ -1,0 +1,26 @@
+import { primary, gray300 } from "@/app/styles/colors.css";
+import { pointer } from "@/app/styles/global.css";
+import Image, { ImageProps } from "next/image";
+import { shapeContentBoxWrapper, shapeContentBox, shapeContentBoxText } from "../styles/shape.css";
+
+type ContentBoxProps = {
+  src: string;
+  text: string;
+  width: ImageProps["width"];
+  height: ImageProps["height"];
+  active?: boolean;
+  onClick: (label: string) => void;
+};
+
+const ContentBox = ({ src, text, width, height, active, onClick }: ContentBoxProps) => {
+  return (
+    <div className={shapeContentBoxWrapper} onClick={() => onClick(text)}>
+      <div className={`${shapeContentBox} ${pointer}`}>
+        <Image src={src} alt={text} width={width} height={height} />
+      </div>
+      <p className={`${shapeContentBoxText} ${active ? primary : gray300}`}>{text}</p>
+    </div>
+  );
+};
+
+export default ContentBox;
